requirements.txt
attrs==22.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cloudpickle==2.2.1
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.38.0
gym==0.25.2
gym-notices==0.0.8
importlib-metadata==6.1.0
iniconfig==2.0.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.5.3
metadrive-simulator==0.3.0.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.0
Panda3D==1.10.13
panda3d-gltf==0.13
panda3d-simplepbr==0.10
pandas==1.3.5
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
psutil==5.9.4
pygame==2.3.0
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
tomli==2.0.1
torch==1.13.1
tqdm==4.65.0
typing_extensions==4.5.0
yapf==0.32.0
zipp==3.15.0
torch-summary==1.4.5
torchviz==0.0.2
